Only send what the page in front of you could really send
Seven things from review, all the same shape: a check that looked at part of
a page and passed on the rest.

- A required box the story never mentioned is now refused. Only the boxes a
  caller named were being looked at, so a form that started insisting on a
  tick nobody made would go through. The setup ceremony was doing exactly
  that: its agreement box was being typed in as a value rather than ticked,
  which is why the new rule catches it.
- A button may say where its own form goes and how it sends. Pressing one is
  now checked against the button's word, not only the form's, so a button
  carrying formaction or formmethod cannot be treated as a post to somewhere
  it would never reach.
- Finding the way into a named thing now fails when the list renders no link,
  rather than answering with nothing and leaving each caller to check. The
  five callers share one helper, and the two that could never answer with
  nothing have lost a message that could not be reached.
- Pressing an arrow the list does not offer fails, instead of quietly doing
  nothing and letting a story claim a row moved.
- The jobs a booking state can hold are one table keyed by a named list, so
  a new job is a missing entry rather than a silent miss.
- The disabled-username example asserted an empty refusal, which accepted any
  error at all. It now names the words it expects.
- test-browser.test.ts is five files under test/test-utils/test-browser/,
  each about one thing, with the stand-ins they share in helpers.ts.

Also fixes a flake this uncovered: two test files were writing the shared
STRIPE_MOCK_PORT while the whole suite ran side by side in one process. The
mutation runner now takes the port as an argument, and the environment test
uses names of its own.
